+++ crates/kitsune_p2p/dht_arc/src/dht_arc.rs
@@ -19,9 +19,6 @@ mod test_ascii;
 mod dht_arc_set;
 pub use dht_arc_set::{ArcInterval, DhtArcSet};
 
-#[cfg(feature = "test_utils")]
-pub(crate) use dht_arc_set::{loc_downscale, loc_upscale};
-
 mod dht_arc_bucket;
 pub use dht_arc_bucket::*;
 
@@ -571,6 +568,37 @@ impl DhtArc {
             }
         }
     }
+}
+
+/// Scale a number in a smaller space (specified by `len`) up into the `u32` space.
+/// The number to scale can be negative, which is wrapped to a positive value via modulo
+#[cfg(any(test, feature = "test_utils"))]
+pub(crate) fn loc_upscale(len: usize, v: i32) -> u32 {
+    let max = 2f64.powi(32);
+    let lenf = len as f64;
+    let vf = v as f64;
+    (max / lenf * vf).round() as i64 as u32
+}
+
+/// Scale a u32 DhtLocation down into a smaller space (specified by `len`)
+#[cfg(any(test, feature = "test_utils"))]
+pub(crate) fn loc_downscale(len: usize, d: DhtLocation) -> usize {
+    let max = 2f64.powi(32);
+    let lenf = len as f64;
+    ((lenf / max * (d.as_u32() as f64)).round() as usize) % len
+}
+
+#[test]
+fn test_loc_upscale() {
+    assert_eq!(loc_upscale(8, 0), DhtLocation::from(0).as_u32());
+    assert_eq!(
+        loc_upscale(8, 1),
+        DhtLocation::from(u32::MAX / 8 + 1).as_u32()
+    );
+    assert_eq!(
+        loc_upscale(3, 1),
+        DhtLocation::from(u32::MAX / 3 + 1).as_u32()
+    );
 }
 
 #[test]
