+++ crates/kitsune_p2p/dht_arc/src/dht_arc/dht_arc_set.rs
@@ -372,6 +372,8 @@ impl ArcInterval<DhtLocation> {
     /// Handy ascii representation of an arc, especially useful when
     /// looking at several arcs at once to get a sense of their overlap
     pub fn to_ascii(&self, len: usize) -> String {
+        use crate::loc_downscale;
+
         match self {
             Self::Full => "-".repeat(len),
             Self::Empty => " ".repeat(len),
@@ -409,7 +411,7 @@ impl ArcInterval<DhtLocation> {
         len: usize,
         ops: I,
     ) -> String {
-        use crate::loc8::Loc8;
+        use crate::{loc8::Loc8, loc_downscale};
 
         let mut buf = vec![0; len];
         let mut s = self.to_ascii(len);
@@ -439,40 +441,9 @@ fn is_full(start: u32, end: u32) -> bool {
     (start == MIN && end >= MAX) || end == start.wrapping_sub(1)
 }
 
-/// Scale a number in a smaller space (specified by `len`) up into the `u32` space.
-/// The number to scale can be negative, which is wrapped to a positive value via modulo
-#[cfg(any(test, feature = "test_utils"))]
-pub(crate) fn loc_upscale(len: usize, v: i32) -> u32 {
-    let max = 2f64.powi(32);
-    let lenf = len as f64;
-    let vf = v as f64;
-    (max / lenf * vf).round() as i64 as u32
-}
-
-/// Scale a u32 DhtLocation down into a smaller space (specified by `len`)
-#[cfg(any(test, feature = "test_utils"))]
-pub(crate) fn loc_downscale(len: usize, d: DhtLocation) -> usize {
-    let max = 2f64.powi(32);
-    let lenf = len as f64;
-    ((lenf / max * (d.as_u32() as f64)).round() as usize) % len
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_loc_upscale() {
-        assert_eq!(loc_upscale(8, 0), DhtLocation::from(0).as_u32());
-        assert_eq!(
-            loc_upscale(8, 1),
-            DhtLocation::from(u32::MAX / 8 + 1).as_u32()
-        );
-        assert_eq!(
-            loc_upscale(3, 1),
-            DhtLocation::from(u32::MAX / 3 + 1).as_u32()
-        );
-    }
 
     #[test]
     fn arc_contains() {
